Code style refactoring
