+++ flink-libraries/flink-table/src/main/scala/org/apache/flink/api/table/plan/nodes/dataset/DataSetSort.scala
@@ -122,6 +122,6 @@ class DataSetSort(
 
   override def explainTerms(pw: RelWriter) : RelWriter = {
     super.explainTerms(pw)
-      .item("by", sortFieldsToString)
+      .item("orderBy", sortFieldsToString)
   }
 }
