+++ flink-libraries/flink-table/src/main/scala/org/apache/flink/api/table/plan/rules/dataSet/DataSetSortRule.scala
@@ -30,7 +30,7 @@ class DataSetSortRule
     classOf[LogicalSort],
     Convention.NONE,
     DataSetConvention.INSTANCE,
-    "FlinkSortRule") {
+    "DataSetSortRule") {
 
   /**
     * Only translate when no OFFSET or LIMIT specified
